make heater swithch on function
1.place the iluminate plane inside the heater
