+++ PhotoTemplate2/Assets/Temperture/Scripts/ShowvalueScript.cs.meta
