+++ PhotoTemplate2/Assets/Temperture/material.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ebe551300c7cd2f4091e472999a9ce0b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
