+++ PhotoTemplate2/Assets/Temperture/material/heater material.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 58e5c5ea853fe8f47a144c9dc5ec4890
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
