+++ PhotoTemplate2/Item/Text.cs
@@ -1,0 +1,6 @@
+﻿namespace Item
+{
+    internal class Text
+    {
+    }
+}
